1040
read 4 float numbers stand for student's scores,
and calculate average with weight 2, 3, 4 and 1
respectively.
If average is between 5.0 to 6.9(included), need to
read another score stand for another exam, and
print original average and new score and show if
it passed or not.
If average is greater or equal then 7.0, show he passed
If average is lower then 5.0 show he did not passed